fix search error, new review list interface, rating fix

+++ flutter_app/lib/models/review.dart
@@ -6,7 +6,7 @@ class Review {
   final int like;
   final String user_id;
   final String image;
-  final double rating;
+  final int rating;
 
   Review({ this.restaurant_id, this.content, this.dislike, this.like, this.user_id, this.image, this.rating});
